pyproject.toml
[tool.poetry]
name = "deepsearch"
version = "0.1.0"
description = ""
authors = ["RupeshBansal"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.220"
pytest = "^7.3.1"
black = "^23.10.1"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
mock = "^5.1.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<3.13"
openai = "^0.28.1"
chromadb = "^0.4.15"
pillow = "10.0.1"
sentence-transformers = "^2.2.2"
torchvision = ">=0.15.1, !=0.15.2"
ftfy = "^6.1.1"
torch = "^2.0.0"
regex = "2023.8.8"
boto3 = "^1.28.72"
openai-whisper = { version = "20230918", optional = true }
streamlit = { version = "^1.28.0", optional = true }
mimetype = "^0.1.5"
pydub = "^0.25.1"
langchain = "^0.0.330"


[tool.poetry.extras]
streamlit = ["streamlit"]
openai-whisper = ["openai-whisper"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
